+++ pkg/util/hardware/mem_info.go
@@ -1,0 +1,43 @@
+//go:build !darwin && !fressbsd && !openbsd
+// +build !darwin,!fressbsd,!openbsd
+
+package hardware
+
+import (
+	"os"
+
+	"github.com/milvus-io/milvus/pkg/log"
+	"github.com/shirou/gopsutil/process"
+	"go.uber.org/zap"
+)
+
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// GetUsedMemoryCount returns the memory usage in bytes.
+func GetUsedMemoryCount() uint64 {
+	proc, err := process.NewProcess(int32(os.Getpid()))
+	if err != nil {
+		log.Warn("failed to get process info", zap.Error(err))
+		return 0
+	}
+
+	memInfo, err := proc.MemoryInfoEx()
+	if err != nil {
+		log.Warn("failed to get memory info", zap.Error(err))
+		return 0
+	}
+
+	// sub the shared memory to filter out the file-backed map memory usage
+	return memInfo.RSS - memInfo.Shared
+}
